Move prefab code to prefab crate
