+++ crates/editor/src/editor_registry.rs
