+++ crates/prefab/src/component/light.rs
