+++ crates/prefab/src/component/material.rs
