+++ crates/prefab/src/component/path.rs
