+++ crates/prefab/src/component/shape.rs
