+++ crates/prefab/src/load.rs
