+++ crates/prefab/src/spawn_system.rs
